+++ backend/src/main/java/com/a407/back/model/service/basicService.java
@@ -11,9 +11,11 @@ import java.util.Optional;
 @Service
 @RequiredArgsConstructor
 public class basicService {
+
     private final basicRepository basicRepository;
+
     @Transactional
-    public int save(User user) {
+    public Long save(User user) {
         return basicRepository.save(user).getId();
     }
 }
